pyproject.toml
[project]
name = "executor-balancer"
version = "0.1.0"
description = "Executor balancer for hakaton"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "celery-types>=0.23.0",
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django>=4.2.25",
    "django-cors-headers>=4.9.0",
    "django-mongoengine>=0.5.6",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "mongoengine>=0.29.1",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
